feat: restructure handlers and implement availability collection system
- Restructure handlers into separate files by trigger/source:
  - handlers/health.go: Basic operational endpoints
  - handlers/scheduled.go: Cloud Scheduler triggered endpoints
  - handlers/pubsub.go: Pub/Sub push subscription endpoints
  - handlers/api.go: REST API endpoints
  - handlers/slack_commands.go: Slack slash command endpoints
  - handlers/slack_events.go: Slack webhook events
  - handlers/helpers.go: Shared helper functions
- Remove old monolithic handlers.go file
- Implement emoji reaction processing for availability collection
- Add database methods for managing player availability
- Fix import cycles and centralize shared constants
- Update server routing to use new handler structure
